+++ test/runtests.jl
@@ -25,6 +25,7 @@ const TEST_MODE = get(ENV, "TEST_MODE", "default")
         include("unit/test_mfa_sampler.jl")
         include("unit/test_exchange.jl")
         include("unit/test_tensorial.jl")
+        include("unit/test_multipole.jl")
     end
     if TEST_MODE in ("default", "all", "aqua")
         include("aqua.jl")